wip: split core simulations
