+++ x/ecocredit/base/simulation/msg_cancel.go
@@ -1,0 +1,98 @@
+package simulation
+
+import (
+	"math/rand"
+
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	simappparams "github.com/cosmos/cosmos-sdk/simapp/params"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
+	"github.com/cosmos/cosmos-sdk/x/simulation"
+
+	"github.com/regen-network/regen-ledger/types/math"
+	"github.com/regen-network/regen-ledger/x/ecocredit"
+	types "github.com/regen-network/regen-ledger/x/ecocredit/base/types/v1"
+	"github.com/regen-network/regen-ledger/x/ecocredit/simulation/utils"
+)
+
+const OpWeightMsgCancel = "op_weight_msg_cancel" //nolint:gosec
+
+const WeightCancel = 30
+
+var TypeMsgCancel = sdk.MsgTypeURL(&types.MsgCancel{})
+
+// SimulateMsgCancel generates a MsgCancel with random values.
+func SimulateMsgCancel(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
+	qryClient types.QueryServer) simtypes.Operation {
+	return func(
+		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
+	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
+
+		ctx := sdk.WrapSDKContext(sdkCtx)
+		class, op, err := utils.GetRandomClass(sdkCtx, r, qryClient, TypeMsgCancel)
+		if class == nil {
+			return op, nil, err
+		}
+
+		project, op, err := getRandomProjectFromClass(ctx, r, qryClient, TypeMsgCancel, class.Id)
+		if project == nil {
+			return op, nil, err
+		}
+
+		batch, op, err := getRandomBatchFromProject(ctx, r, qryClient, TypeMsgCancel, project.Id)
+		if batch == nil {
+			return op, nil, err
+		}
+
+		admin := sdk.AccAddress(project.Admin).String()
+		balanceRes, err := qryClient.Balance(ctx, &types.QueryBalanceRequest{
+			Address:    admin,
+			BatchDenom: batch.Denom,
+		})
+		if err != nil {
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCancel, err.Error()), nil, err
+		}
+
+		tradableBalance, err := math.NewNonNegativeDecFromString(balanceRes.Balance.TradableAmount)
+		if err != nil {
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCancel, err.Error()), nil, err
+		}
+
+		if tradableBalance.IsZero() {
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCancel, "balance is zero"), nil, nil
+		}
+
+		msg := &types.MsgCancel{
+			Owner: admin,
+			Credits: []*types.Credits{
+				{
+					BatchDenom: batch.Denom,
+					Amount:     balanceRes.Balance.TradableAmount,
+				},
+			},
+			Reason: simtypes.RandStringOfLength(r, 5),
+		}
+
+		spendable, account, op, err := utils.GetAccountAndSpendableCoins(sdkCtx, bk, accs, admin, TypeMsgCancel)
+		if spendable == nil {
+			return op, nil, err
+		}
+
+		txCtx := simulation.OperationInput{
+			R:               r,
+			App:             app,
+			TxGen:           simappparams.MakeTestEncodingConfig().TxConfig,
+			Cdc:             nil,
+			Msg:             msg,
+			MsgType:         msg.Type(),
+			Context:         sdkCtx,
+			SimAccount:      *account,
+			AccountKeeper:   ak,
+			Bankkeeper:      bk,
+			ModuleName:      ecocredit.ModuleName,
+			CoinsSpentInMsg: spendable,
+		}
+
+		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
+	}
+}
