+++ x/ecocredit/base/simulation/operations.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"math/rand"
 	"strings"
-	"time"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -31,15 +30,9 @@ import (
 
 // Simulation operation weights constants
 const (
-	OpWeightMsgCreateClass              = "op_weight_msg_create_class"                //nolint:gosec
-	OpWeightMsgCreateBatch              = "op_weight_msg_create_batch"                //nolint:gosec
-	OpWeightMsgSend                     = "op_weight_msg_send"                        //nolint:gosec
-	OpWeightMsgRetire                   = "op_weight_msg_retire"                      //nolint:gosec
-	OpWeightMsgCancel                   = "op_weight_msg_cancel"                      //nolint:gosec
 	OpWeightMsgUpdateClassAdmin         = "op_weight_msg_update_class_admin"          //nolint:gosec
 	OpWeightMsgUpdateClassMetadata      = "op_weight_msg_update_class_metadata"       //nolint:gosec
 	OpWeightMsgUpdateClassIssuers       = "op_weight_msg_update_class_issuers"        //nolint:gosec
-	OpWeightMsgCreateProject            = "op_weight_msg_create_project"              //nolint:gosec
 	OpWeightMsgUpdateProjectAdmin       = "op_weight_msg_update_project_admin"        //nolint:gosec
 	OpWeightMsgUpdateProjectMetadata    = "op_weight_msg_update_project_metadata"     //nolint:gosec
 	OpWeightMsgMintBatchCredits         = "op_weight_msg_mint_batch_credits"          //nolint:gosec
@@ -54,12 +47,6 @@ const (
 
 // ecocredit operations weights
 const (
-	WeightCreateClass           = 10
-	WeightCreateProject         = 20
-	WeightCreateBatch           = 50
-	WeightSend                  = 100
-	WeightRetire                = 80
-	WeightCancel                = 30
 	WeightUpdateClass           = 30
 	WeightUpdateProjectAdmin    = 30
 	WeightUpdateProjectMetadata = 30
@@ -70,12 +57,6 @@ const (
 
 // ecocredit message types
 var (
-	TypeMsgCreateClass              = sdk.MsgTypeURL(&types.MsgCreateClass{})
-	TypeMsgCreateProject            = sdk.MsgTypeURL(&types.MsgCreateProject{})
-	TypeMsgCreateBatch              = sdk.MsgTypeURL(&types.MsgCreateBatch{})
-	TypeMsgSend                     = sdk.MsgTypeURL(&types.MsgSend{})
-	TypeMsgRetire                   = sdk.MsgTypeURL(&types.MsgRetire{})
-	TypeMsgCancel                   = sdk.MsgTypeURL(&types.MsgCancel{})
 	TypeMsgUpdateClassAdmin         = sdk.MsgTypeURL(&types.MsgUpdateClassAdmin{})
 	TypeMsgUpdateClassIssuers       = sdk.MsgTypeURL(&types.MsgUpdateClassIssuers{})
 	TypeMsgUpdateClassMetadata      = sdk.MsgTypeURL(&types.MsgUpdateClassMetadata{})
@@ -404,457 +385,6 @@ func SimulateMsgUpdateProjectAdmin(ak ecocredit.AccountKeeper, bk ecocredit.Bank
 		}
 
 		spendable, account, op, err := utils.GetAccountAndSpendableCoins(sdkCtx, bk, accs, project.Admin, TypeMsgUpdateProjectAdmin)
-		if spendable == nil {
-			return op, nil, err
-		}
-
-		txCtx := simulation.OperationInput{
-			R:               r,
-			App:             app,
-			TxGen:           simappparams.MakeTestEncodingConfig().TxConfig,
-			Cdc:             nil,
-			Msg:             msg,
-			MsgType:         msg.Type(),
-			Context:         sdkCtx,
-			SimAccount:      *account,
-			AccountKeeper:   ak,
-			Bankkeeper:      bk,
-			ModuleName:      ecocredit.ModuleName,
-			CoinsSpentInMsg: spendable,
-		}
-
-		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
-	}
-}
-
-// SimulateMsgCreateClass generates a MsgCreateClass with random values.
-func SimulateMsgCreateClass(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	qryClient types.QueryServer) simtypes.Operation {
-	return func(
-		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-		admin, _ := simtypes.RandomAcc(r, accs)
-		issuers := randomIssuers(r, accs)
-
-		ctx := sdk.WrapSDKContext(sdkCtx)
-		res, err := qryClient.Params(ctx, &types.QueryParamsRequest{})
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCreateClass, err.Error()), nil, err
-		}
-
-		params := res.Params
-		if params.AllowlistEnabled && !utils.Contains(params.AllowedClassCreators, admin.Address.String()) {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCreateClass, "not allowed to create credit class"), nil, nil // skip
-		}
-
-		spendable, neg := bk.SpendableCoins(sdkCtx, admin.Address).SafeSub(params.CreditClassFee...)
-		if neg {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCreateClass, "not enough balance"), nil, nil
-		}
-
-		creditTypes := []string{"C", "BIO"}
-		msg := &types.MsgCreateClass{
-			Admin:            admin.Address.String(),
-			Issuers:          issuers,
-			Metadata:         simtypes.RandStringOfLength(r, 10),
-			CreditTypeAbbrev: creditTypes[r.Intn(len(creditTypes))],
-			Fee:              &params.CreditClassFee[0],
-		}
-
-		txCtx := simulation.OperationInput{
-			R:               r,
-			App:             app,
-			TxGen:           simappparams.MakeTestEncodingConfig().TxConfig,
-			Cdc:             nil,
-			Msg:             msg,
-			MsgType:         msg.Type(),
-			Context:         sdkCtx,
-			SimAccount:      admin,
-			AccountKeeper:   ak,
-			Bankkeeper:      bk,
-			ModuleName:      ecocredit.ModuleName,
-			CoinsSpentInMsg: spendable,
-		}
-
-		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
-	}
-}
-
-// SimulateMsgCreateProject generates a MsgCreateProject with random values.
-func SimulateMsgCreateProject(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	qryClient types.QueryServer) simtypes.Operation {
-	return func(
-		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-		class, op, err := utils.GetRandomClass(sdkCtx, r, qryClient, TypeMsgCreateProject)
-		if class == nil {
-			return op, nil, err
-		}
-
-		issuers, op, err := getClassIssuers(sdkCtx, qryClient, class.Id, TypeMsgCreateProject)
-		if len(issuers) == 0 {
-			return op, nil, err
-		}
-
-		admin := issuers[r.Intn(len(issuers))]
-		adminAddr, err := sdk.AccAddressFromBech32(admin)
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCreateProject, err.Error()), nil, err
-		}
-
-		adminAcc, found := simtypes.FindAccount(accs, adminAddr)
-		if !found {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCreateProject, "not a simulation account"), nil, nil
-		}
-
-		spendable := bk.SpendableCoins(sdkCtx, adminAddr)
-
-		msg := &types.MsgCreateProject{
-			Admin:        admin,
-			ClassId:      class.Id,
-			Metadata:     simtypes.RandStringOfLength(r, 100),
-			Jurisdiction: "AB-CDE FG1 345",
-		}
-		txCtx := simulation.OperationInput{
-			R:               r,
-			App:             app,
-			TxGen:           simappparams.MakeTestEncodingConfig().TxConfig,
-			Cdc:             nil,
-			Msg:             msg,
-			MsgType:         msg.Type(),
-			Context:         sdkCtx,
-			SimAccount:      adminAcc,
-			AccountKeeper:   ak,
-			Bankkeeper:      bk,
-			ModuleName:      ecocredit.ModuleName,
-			CoinsSpentInMsg: spendable,
-		}
-
-		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
-	}
-}
-
-// SimulateMsgCreateBatch generates a MsgCreateBatch with random values.
-func SimulateMsgCreateBatch(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	qryClient types.QueryServer) simtypes.Operation {
-	return func(
-		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-		issuer, _ := simtypes.RandomAcc(r, accs)
-
-		ctx := sdk.WrapSDKContext(sdkCtx)
-		class, op, err := utils.GetRandomClass(sdkCtx, r, qryClient, TypeMsgCreateBatch)
-		if class == nil {
-			return op, nil, err
-		}
-
-		result, err := qryClient.ClassIssuers(ctx, &types.QueryClassIssuersRequest{ClassId: class.Id})
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCreateBatch, err.Error()), nil, err
-		}
-
-		classIssuers := result.Issuers
-		if len(classIssuers) == 0 {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCreateBatch, "no issuers"), nil, nil
-		}
-
-		project, op, err := getRandomProjectFromClass(ctx, r, qryClient, TypeMsgCreateBatch, class.Id)
-		if project == nil {
-			return op, nil, err
-		}
-
-		if !utils.Contains(classIssuers, issuer.Address.String()) {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCreateBatch, "don't have permission to create credit batch"), nil, nil
-		}
-
-		issuerAcc := ak.GetAccount(sdkCtx, issuer.Address)
-		spendable := bk.SpendableCoins(sdkCtx, issuerAcc.GetAddress())
-
-		now := sdkCtx.BlockTime()
-		tenHours := now.Add(10 * time.Hour)
-
-		msg := &types.MsgCreateBatch{
-			Issuer:    issuer.Address.String(),
-			ProjectId: project.Id,
-			Issuance:  generateBatchIssuance(r, accs),
-			StartDate: &now,
-			EndDate:   &tenHours,
-			Metadata:  simtypes.RandStringOfLength(r, 10),
-			Open:      r.Float32() < 0.3, // 30% chance of credit batch being dynamic batch
-		}
-
-		txCtx := simulation.OperationInput{
-			R:               r,
-			App:             app,
-			TxGen:           simappparams.MakeTestEncodingConfig().TxConfig,
-			Cdc:             nil,
-			Msg:             msg,
-			MsgType:         msg.Type(),
-			Context:         sdkCtx,
-			SimAccount:      issuer,
-			AccountKeeper:   ak,
-			Bankkeeper:      bk,
-			ModuleName:      ecocredit.ModuleName,
-			CoinsSpentInMsg: spendable,
-		}
-
-		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
-	}
-}
-
-// SimulateMsgSend generates a MsgSend with random values.
-func SimulateMsgSend(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	qryClient types.QueryServer) simtypes.Operation {
-	return func(
-		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-
-		ctx := sdk.WrapSDKContext(sdkCtx)
-		class, op, err := utils.GetRandomClass(sdkCtx, r, qryClient, TypeMsgSend)
-		if class == nil {
-			return op, nil, err
-		}
-
-		project, op, err := getRandomProjectFromClass(ctx, r, qryClient, TypeMsgSend, class.Id)
-		if project == nil {
-			return op, nil, err
-		}
-
-		batch, op, err := getRandomBatchFromProject(ctx, r, qryClient, TypeMsgSend, class.Id)
-		if batch == nil {
-			return op, nil, err
-		}
-
-		admin := sdk.AccAddress(project.Admin).String()
-		balres, err := qryClient.Balance(ctx, &types.QueryBalanceRequest{
-			Address:    admin,
-			BatchDenom: batch.Denom,
-		})
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSend, err.Error()), nil, err
-		}
-
-		tradableBalance, err := math.NewNonNegativeDecFromString(balres.Balance.TradableAmount)
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSend, err.Error()), nil, err
-		}
-
-		retiredBalance, err := math.NewNonNegativeDecFromString(balres.Balance.RetiredAmount)
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSend, err.Error()), nil, err
-		}
-
-		if tradableBalance.IsZero() || retiredBalance.IsZero() {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSend, "balance is zero"), nil, nil
-		}
-
-		recipient, _ := simtypes.RandomAcc(r, accs)
-		if admin == recipient.Address.String() {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSend, "sender & recipient are same"), nil, nil
-		}
-
-		addr := sdk.AccAddress(project.Admin)
-		acc, found := simtypes.FindAccount(accs, addr)
-		if !found {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSend, "account not found"), nil, nil
-		}
-
-		issuer := ak.GetAccount(sdkCtx, acc.Address)
-		spendable := bk.SpendableCoins(sdkCtx, issuer.GetAddress())
-
-		var tradable int
-		var retired int
-		var retirementJurisdiction string
-		if !tradableBalance.IsZero() {
-			i64, err := tradableBalance.Int64()
-			if err != nil {
-				return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSend, err.Error()), nil, nil
-			}
-			if i64 > 1 {
-				tradable = simtypes.RandIntBetween(r, 1, int(i64))
-				retired = simtypes.RandIntBetween(r, 0, tradable)
-				if retired != 0 {
-					retirementJurisdiction = "AQ"
-				}
-			} else {
-				tradable = int(i64)
-			}
-		}
-
-		if retired+tradable > tradable {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSend, "insufficient credit balance"), nil, nil
-		}
-
-		msg := &types.MsgSend{
-			Sender:    admin,
-			Recipient: recipient.Address.String(),
-			Credits: []*types.MsgSend_SendCredits{
-				{
-					BatchDenom:             batch.Denom,
-					TradableAmount:         fmt.Sprintf("%d", tradable),
-					RetiredAmount:          fmt.Sprintf("%d", retired),
-					RetirementJurisdiction: retirementJurisdiction,
-				},
-			},
-		}
-
-		txCtx := simulation.OperationInput{
-			R:               r,
-			App:             app,
-			TxGen:           simappparams.MakeTestEncodingConfig().TxConfig,
-			Cdc:             nil,
-			Msg:             msg,
-			MsgType:         msg.Type(),
-			Context:         sdkCtx,
-			SimAccount:      acc,
-			AccountKeeper:   ak,
-			Bankkeeper:      bk,
-			ModuleName:      ecocredit.ModuleName,
-			CoinsSpentInMsg: spendable,
-		}
-
-		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
-	}
-}
-
-// SimulateMsgRetire generates a MsgRetire with random values.
-func SimulateMsgRetire(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	qryClient types.QueryServer) simtypes.Operation {
-	return func(
-		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-
-		ctx := sdk.WrapSDKContext(sdkCtx)
-		class, op, err := utils.GetRandomClass(sdkCtx, r, qryClient, TypeMsgRetire)
-		if class == nil {
-			return op, nil, err
-		}
-
-		project, op, err := getRandomProjectFromClass(ctx, r, qryClient, TypeMsgRetire, class.Id)
-		if project == nil {
-			return op, nil, err
-		}
-
-		batch, op, err := getRandomBatchFromProject(ctx, r, qryClient, TypeMsgRetire, project.Id)
-		if batch == nil {
-			return op, nil, err
-		}
-
-		admin := sdk.AccAddress(project.Admin).String()
-		balanceRes, err := qryClient.Balance(ctx, &types.QueryBalanceRequest{
-			Address:    admin,
-			BatchDenom: batch.Denom,
-		})
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRetire, err.Error()), nil, err
-		}
-
-		tradableBalance, err := math.NewNonNegativeDecFromString(balanceRes.Balance.TradableAmount)
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRetire, err.Error()), nil, err
-		}
-
-		if tradableBalance.IsZero() {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRetire, "balance is zero"), nil, nil
-		}
-
-		randSub := math.NewDecFromInt64(int64(simtypes.RandIntBetween(r, 1, 10)))
-		spendable, account, op, err := utils.GetAccountAndSpendableCoins(sdkCtx, bk, accs, admin, TypeMsgRetire)
-		if spendable == nil {
-			return op, nil, err
-		}
-
-		if !spendable.IsAllPositive() {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRetire, "insufficient funds"), nil, nil
-		}
-
-		if tradableBalance.Cmp(randSub) != 1 {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRetire, "insufficient funds"), nil, nil
-		}
-
-		msg := &types.MsgRetire{
-			Owner: account.Address.String(),
-			Credits: []*types.Credits{
-				{
-					BatchDenom: batch.Denom,
-					Amount:     randSub.String(),
-				},
-			},
-			Jurisdiction: "ST-UVW XY Z12",
-		}
-
-		txCtx := simulation.OperationInput{
-			R:               r,
-			App:             app,
-			TxGen:           simappparams.MakeTestEncodingConfig().TxConfig,
-			Cdc:             nil,
-			Msg:             msg,
-			MsgType:         msg.Type(),
-			Context:         sdkCtx,
-			SimAccount:      *account,
-			AccountKeeper:   ak,
-			Bankkeeper:      bk,
-			ModuleName:      ecocredit.ModuleName,
-			CoinsSpentInMsg: spendable,
-		}
-
-		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
-	}
-}
-
-// SimulateMsgCancel generates a MsgCancel with random values.
-func SimulateMsgCancel(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	qryClient types.QueryServer) simtypes.Operation {
-	return func(
-		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-
-		ctx := sdk.WrapSDKContext(sdkCtx)
-		class, op, err := utils.GetRandomClass(sdkCtx, r, qryClient, TypeMsgCancel)
-		if class == nil {
-			return op, nil, err
-		}
-
-		project, op, err := getRandomProjectFromClass(ctx, r, qryClient, TypeMsgCancel, class.Id)
-		if project == nil {
-			return op, nil, err
-		}
-
-		batch, op, err := getRandomBatchFromProject(ctx, r, qryClient, TypeMsgCancel, project.Id)
-		if batch == nil {
-			return op, nil, err
-		}
-
-		admin := sdk.AccAddress(project.Admin).String()
-		balanceRes, err := qryClient.Balance(ctx, &types.QueryBalanceRequest{
-			Address:    admin,
-			BatchDenom: batch.Denom,
-		})
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCancel, err.Error()), nil, err
-		}
-
-		tradableBalance, err := math.NewNonNegativeDecFromString(balanceRes.Balance.TradableAmount)
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCancel, err.Error()), nil, err
-		}
-
-		if tradableBalance.IsZero() {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgCancel, "balance is zero"), nil, nil
-		}
-
-		msg := &types.MsgCancel{
-			Owner: admin,
-			Credits: []*types.Credits{
-				{
-					BatchDenom: batch.Denom,
-					Amount:     balanceRes.Balance.TradableAmount,
-				},
-			},
-			Reason: simtypes.RandStringOfLength(r, 5),
-		}
-
-		spendable, account, op, err := utils.GetAccountAndSpendableCoins(sdkCtx, bk, accs, admin, TypeMsgCancel)
 		if spendable == nil {
 			return op, nil, err
 		}
