+++ x/ecocredit/marketplace/simulation/operations.go
@@ -1,40 +1,16 @@
 package simulation
 
 import (
-	"fmt"
 	"math/rand"
 
-	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/codec"
-	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	"github.com/cosmos/cosmos-sdk/simapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
 
 	"github.com/regen-network/regen-ledger/x/ecocredit"
 	basetypes "github.com/regen-network/regen-ledger/x/ecocredit/base/types/v1"
 	types "github.com/regen-network/regen-ledger/x/ecocredit/marketplace/types/v1"
-	"github.com/regen-network/regen-ledger/x/ecocredit/simulation/utils"
-)
-
-// Simulation operation weights constants
-const (
-	OpWeightMsgAddAllowedDenom    = "op_weight_msg_add_allowed_denom"    //nolint:gosec
-	OpWeightMsgRemoveAllowedDenom = "op_weight_msg_remove_allowed_denom" //nolint:gosec
-)
-
-// basket operations weights
-const (
-	WeightAddAllowedDenom    = 100
-	WeightRemoveAllowedDenom = 100
-)
-
-// ecocredit message types
-var (
-	TypeMsgAddAllowedDenom    = types.MsgAddAllowedDenom{}.Route()
-	TypeMsgRemoveAllowedDenom = types.MsgRemoveAllowedDenom{}.Route()
 )
 
 func WeightedOperations(
@@ -113,154 +89,5 @@ func WeightedOperations(
 			weightMsgRemoveAllowedDenom,
 			SimulateMsgRemoveAllowedDenom(ak, bk, mktQryClient, govk, authority),
 		),
-	}
-}
-
-func SimulateMsgAddAllowedDenom(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	qryClient types.QueryServer, govk ecocredit.GovKeeper, authority sdk.AccAddress) simtypes.Operation {
-	return func(
-		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-
-		proposer, _ := simtypes.RandomAcc(r, accs)
-		proposerAddr := proposer.Address.String()
-
-		spendable, account, op, err := utils.GetAccountAndSpendableCoins(sdkCtx, bk, accs, proposerAddr, TypeMsgAddAllowedDenom)
-		if spendable == nil {
-			return op, nil, err
-		}
-
-		bankDenom := simtypes.RandStringOfLength(r, 4)
-		res, err := qryClient.AllowedDenoms(sdkCtx, &types.QueryAllowedDenomsRequest{})
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgAddAllowedDenom, err.Error()), nil, err
-		}
-
-		if isDenomExists(res.AllowedDenoms, bankDenom) {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgAddAllowedDenom, fmt.Sprintf("denom %s already exists", bankDenom)), nil, nil
-		}
-
-		params := govk.GetDepositParams(sdkCtx)
-		deposit, skip, err := utils.RandomDeposit(r, sdkCtx, ak, bk, params, authority)
-		switch {
-		case skip:
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgAddAllowedDenom, "skip deposit"), nil, nil
-		case err != nil:
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgAddAllowedDenom, "unable to generate deposit"), nil, err
-		}
-
-		msg := types.MsgAddAllowedDenom{
-			Authority:    authority.String(),
-			BankDenom:    bankDenom,
-			DisplayDenom: bankDenom,
-			Exponent:     6,
-		}
-
-		any, err := codectypes.NewAnyWithValue(&msg)
-		if err != nil {
-			return simtypes.NoOpMsg(TypeMsgAddAllowedDenom, TypeMsgAddAllowedDenom, err.Error()), nil, err
-		}
-
-		proposalMsg := govtypes.MsgSubmitProposal{
-			InitialDeposit: deposit,
-			Proposer:       proposerAddr,
-			Metadata:       simtypes.RandStringOfLength(r, 10),
-			Messages:       []*codectypes.Any{any},
-		}
-
-		txCtx := simulation.OperationInput{
-			R:               r,
-			App:             app,
-			TxGen:           simapp.MakeTestEncodingConfig().TxConfig,
-			Cdc:             nil,
-			Msg:             &proposalMsg,
-			MsgType:         msg.Type(),
-			Context:         sdkCtx,
-			SimAccount:      *account,
-			AccountKeeper:   ak,
-			Bankkeeper:      bk,
-			ModuleName:      ecocredit.ModuleName,
-			CoinsSpentInMsg: spendable,
-		}
-
-		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
-	}
-}
-
-func isDenomExists(allowedDenom []*types.AllowedDenom, bankDenom string) bool {
-	for _, denom := range allowedDenom {
-		if denom.BankDenom == bankDenom {
-			return true
-		}
-	}
-
-	return false
-}
-
-func SimulateMsgRemoveAllowedDenom(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
-	mktClient types.QueryServer, govk ecocredit.GovKeeper, authority sdk.AccAddress) simtypes.Operation {
-	return func(
-		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-
-		response, err := mktClient.AllowedDenoms(sdkCtx, &types.QueryAllowedDenomsRequest{})
-		if err != nil {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRemoveAllowedDenom, err.Error()), nil, err
-		}
-
-		if len(response.AllowedDenoms) == 0 {
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRemoveAllowedDenom, "no allowed denom present"), nil, nil
-		}
-
-		proposer, _ := simtypes.RandomAcc(r, accs)
-		proposerAddr := proposer.Address.String()
-
-		spendable, account, op, err := utils.GetAccountAndSpendableCoins(sdkCtx, bk, accs, proposerAddr, TypeMsgRemoveAllowedDenom)
-		if spendable == nil {
-			return op, nil, err
-		}
-
-		params := govk.GetDepositParams(sdkCtx)
-		deposit, skip, err := utils.RandomDeposit(r, sdkCtx, ak, bk, params, authority)
-		switch {
-		case skip:
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRemoveAllowedDenom, "skip deposit"), nil, nil
-		case err != nil:
-			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgRemoveAllowedDenom, "unable to generate deposit"), nil, err
-		}
-
-		msg := types.MsgRemoveAllowedDenom{
-			Authority: authority.String(),
-			Denom:     response.AllowedDenoms[r.Intn(len(response.AllowedDenoms))].BankDenom,
-		}
-
-		any, err := codectypes.NewAnyWithValue(&msg)
-		if err != nil {
-			return simtypes.NoOpMsg(TypeMsgAddAllowedDenom, TypeMsgAddAllowedDenom, err.Error()), nil, err
-		}
-
-		proposalMsg := govtypes.MsgSubmitProposal{
-			InitialDeposit: deposit,
-			Proposer:       proposerAddr,
-			Metadata:       simtypes.RandStringOfLength(r, 10),
-			Messages:       []*codectypes.Any{any},
-		}
-
-		txCtx := simulation.OperationInput{
-			R:               r,
-			App:             app,
-			TxGen:           simapp.MakeTestEncodingConfig().TxConfig,
-			Cdc:             nil,
-			Msg:             &proposalMsg,
-			MsgType:         msg.Type(),
-			Context:         sdkCtx,
-			SimAccount:      *account,
-			AccountKeeper:   ak,
-			Bankkeeper:      bk,
-			ModuleName:      ecocredit.ModuleName,
-			CoinsSpentInMsg: spendable,
-		}
-
-		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
 	}
 }
